refactor canon
