+++ canon/src/main/scala/cats/canon/FunctorFromApplicative.scala
@@ -1,0 +1,7 @@
+package cats
+package canon
+
+trait FunctorFromApplicative[F[_]] extends Functor[F] { self: Applicative[F] =>
+  def map[A, B](fa: F[A])(f: A => B): F[B] = ap(fa)(pure(f))
+}
+
