+++ tests/src/test/scala/cats/tests/RegressionTests.scala
@@ -1,7 +1,7 @@
 package cats
 package tests
 
-import cats.canon.{CanonicalApplicativeFromMonad}
+import cats.canon.{ApplicativeFromMonad}
 
 import scala.collection.mutable
 
@@ -18,7 +18,7 @@ class RegressionTests extends CatsSuite {
 
   object State {
     implicit def instance[S]: Monad[State[S, ?]] = new Monad[State[S, ?]]
-        with CanonicalApplicativeFromMonad[State[S, ?]] {
+        with ApplicativeFromMonad[State[S, ?]] {
       def pure[A](a: A): State[S, A] = State(s => (a, s))
       override def map[A, B](sa: State[S, A])(f: A => B): State[S, B] = sa.map(f)
       def flatMap[A, B](sa: State[S, A])(f: A => State[S, B]): State[S, B] = sa.flatMap(f)
